split read tx compatibility functions out of the big module
