+++ lib/wallet/cardano-wallet.cabal
@@ -361,6 +361,13 @@ library
       Cardano.Wallet.TokenMetadata.MockServer
       Cardano.Wallet.Transaction
       Cardano.Wallet.Types.Read
+      Cardano.Wallet.Types.Read.Primitive.Tx
+      Cardano.Wallet.Types.Read.Primitive.Tx.Allegra
+      Cardano.Wallet.Types.Read.Primitive.Tx.Alonzo
+      Cardano.Wallet.Types.Read.Primitive.Tx.Babbage
+      Cardano.Wallet.Types.Read.Primitive.Tx.Byron
+      Cardano.Wallet.Types.Read.Primitive.Tx.Mary
+      Cardano.Wallet.Types.Read.Primitive.Tx.Shelley
       Cardano.Wallet.Types.Read.Tx
       Cardano.Wallet.Types.Read.Tx.CBOR
       Cardano.Wallet.Types.Read.Tx.Hash
