+++ lib/wallet/src/Cardano/Wallet/Types/Read/Primitive/Tx/Babbage.hs
@@ -1,0 +1,170 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TupleSections #-}
+{-# LANGUAGE TypeFamilies #-}
+
+-- |
+-- Copyright: © 2020 IOHK
+-- License: Apache-2.0
+--
+-- Conversion functions and static chain settings for Shelley.
+module Cardano.Wallet.Types.Read.Primitive.Tx.Babbage (fromBabbageTx)
+ where
+
+import Prelude
+
+import Cardano.Address.Script
+    ( KeyRole (..) )
+import Cardano.Api
+    ( BabbageEra, CardanoEra (..) )
+import Cardano.Ledger.Era
+    ( Era (..) )
+import Cardano.Ledger.Serialization
+    ( sizedValue )
+import Cardano.Ledger.Shelley.API
+    ( StrictMaybe (SJust, SNothing) )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId )
+import Cardano.Wallet.Shelley.Compatibility.Ledger
+    ( toWalletScript, toWalletTokenPolicyId )
+import Cardano.Wallet.Transaction
+    ( AnyScript (..)
+    , PlutusScriptInfo (..)
+    , PlutusVersion (..)
+    , TokenMapWithScripts (..)
+    , ValidityIntervalExplicit (..)
+    )
+import Cardano.Wallet.Types.Read.Primitive.Tx.Allegra
+    ( fromLedgerTxValidity )
+import Cardano.Wallet.Types.Read.Primitive.Tx.Alonzo
+    ( alonzoTxHash )
+import Cardano.Wallet.Types.Read.Primitive.Tx.Mary
+    ( fromCardanoValue, fromLedgerMintValue, getScriptMap )
+import Cardano.Wallet.Types.Read.Primitive.Tx.Shelley
+    ( fromShelleyAddress
+    , fromShelleyCert
+    , fromShelleyCoin
+    , fromShelleyMD
+    , fromShelleyTxIn
+    , fromShelleyWdrl
+    )
+import Cardano.Wallet.Types.Read.Tx
+    ( Tx (..) )
+import Cardano.Wallet.Types.Read.Tx.CBOR
+    ( getTxCBOR )
+import Data.Foldable
+    ( toList )
+import Data.Map.Strict
+    ( Map )
+import Ouroboros.Consensus.Cardano.Block
+    ( StandardBabbage )
+
+import qualified Cardano.Api.Shelley as Cardano
+import qualified Cardano.Ledger.Alonzo.Data as Alonzo
+import qualified Cardano.Ledger.Alonzo.Language as Alonzo
+import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
+import qualified Cardano.Ledger.Alonzo.Tx as Alonzo
+import qualified Cardano.Ledger.Babbage as Babbage
+import qualified Cardano.Ledger.Babbage.TxBody as Babbage
+import qualified Cardano.Ledger.BaseTypes as SL
+import qualified Cardano.Ledger.Core as SL.Core
+import qualified Cardano.Ledger.Mary.Value as SL
+import qualified Cardano.Ledger.Shelley.API as SL
+import qualified Cardano.Wallet.Primitive.Types as W
+import qualified Cardano.Wallet.Primitive.Types.Coin as W
+import qualified Cardano.Wallet.Primitive.Types.Tx as W
+import qualified Data.Map.Strict as Map
+
+fromBabbageTx
+    :: Alonzo.ValidatedTx (Cardano.ShelleyLedgerEra BabbageEra)
+    -> ( W.Tx
+       , [W.Certificate]
+       , TokenMapWithScripts
+       , TokenMapWithScripts
+       , Maybe ValidityIntervalExplicit
+       )
+fromBabbageTx tx@(Alonzo.ValidatedTx bod wits (Alonzo.IsValid isValid) aux) =
+    ( W.Tx
+        { txId =
+            alonzoTxHash tx
+        , txCBOR =
+            Just $ getTxCBOR $ Tx BabbageEra tx
+        , fee =
+            Just $ fromShelleyCoin fee
+        , resolvedInputs =
+            map ((,W.Coin 0) . fromShelleyTxIn) (toList inps)
+        , resolvedCollateralInputs =
+            map ((,W.Coin 0) . fromShelleyTxIn) (toList collateralInps)
+        , outputs =
+            map (fromBabbageTxOut . sizedValue) (toList outs)
+        , collateralOutput =
+            case fmap (fromBabbageTxOut . sizedValue) collateralReturn of
+                SNothing -> Nothing
+                SJust txout -> Just txout
+        , withdrawals =
+            fromShelleyWdrl wdrls
+        , metadata =
+            fromShelleyMD . toSLMetadata <$> SL.strictMaybeToMaybe aux
+        , scriptValidity =
+            validity
+        }
+    , map fromShelleyCert (toList certs)
+    , TokenMapWithScripts assetsToMint mintScriptMap
+    , TokenMapWithScripts assetsToBurn burnScriptMap
+    , Just (fromLedgerTxValidity ttl)
+    )
+  where
+    Babbage.TxBody
+        inps
+        collateralInps
+        _refInps
+        outs
+        collateralReturn
+        _collateralTotal
+        certs
+        wdrls
+        fee
+        ttl
+        _upd
+        _reqSignerHashes
+        mint
+        _wwpHash
+        _adHash
+        _network
+        = bod
+    (assetsToMint, assetsToBurn) = fromLedgerMintValue mint
+    scriptMap = fromBabbageScriptMap $ Alonzo.txscripts' wits
+    mintScriptMap = getScriptMap scriptMap assetsToMint
+    burnScriptMap = getScriptMap scriptMap assetsToBurn
+
+    fromBabbageScriptMap
+        :: Map
+            (SL.ScriptHash (Crypto StandardBabbage))
+            (SL.Core.Script StandardBabbage)
+        -> Map TokenPolicyId AnyScript
+    fromBabbageScriptMap =
+        Map.map toAnyScript .
+        Map.mapKeys (toWalletTokenPolicyId . SL.PolicyID)
+      where
+        toAnyScript (Alonzo.TimelockScript script) =
+            NativeScript $ toWalletScript Policy script
+        toAnyScript (Alonzo.PlutusScript ver _) =
+            PlutusScript (PlutusScriptInfo (toPlutusVer ver))
+
+        toPlutusVer Alonzo.PlutusV1 = PlutusVersionV1
+        toPlutusVer Alonzo.PlutusV2 = PlutusVersionV2
+
+    fromBabbageTxOut
+        :: Babbage.TxOut (Cardano.ShelleyLedgerEra BabbageEra)
+        -> W.TxOut
+    fromBabbageTxOut (Babbage.TxOut addr value _datum _refScript) =
+        W.TxOut (fromShelleyAddress addr) $
+        fromCardanoValue $ Cardano.fromMaryValue value
+
+    toSLMetadata (Alonzo.AuxiliaryData blob _scripts) = SL.Metadata blob
+
+    validity =
+        if isValid
+        then Just W.TxScriptValid
+        else Just W.TxScriptInvalid
